backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.0
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
eventlet==0.33.3